glsl: work with integer to compute color instead of float
GS likely uses integer to do color operation. It impacts the rounding of color.
For example 129.9 on GS is 129 whereas GPU will round it up to 130.

Integer might be a bit slower but at least it makes the code more
readable for masking operation

Suikoden behaves better with this code. SotC is still bad (over-bright)